moved visitor class from compares module to operators module
