+++ tests/test_visitors/test_ast/test_operators/test_bitwise_boolean_mixup.py
@@ -3,7 +3,7 @@ import pytest
 from wemake_python_styleguide.violations.best_practices import (
     BitwiseAndBooleanMixupViolation,
 )
-from wemake_python_styleguide.visitors.ast.compares import BitwiseOpVisitor
+from wemake_python_styleguide.visitors.ast.operators import BitwiseOpVisitor
 
 # add more cases for more usages of mixup
