+++ wemake_python_styleguide/presets/types/tree.py
@@ -77,6 +77,7 @@ PRESET: Final = (
     operators.UselessOperatorsVisitor,
     operators.WrongMathOperatorVisitor,
     operators.WalrusVisitor,
+    operators.BitwiseOpVisitor,
 
     compares.WrongConditionalVisitor,
     compares.CompareSanityVisitor,
@@ -85,7 +86,6 @@ PRESET: Final = (
     compares.WrongConstantCompareVisitor,
     compares.InCompareSanityVisitor,
     compares.WrongFloatComplexCompareVisitor,
-    compares.BitwiseOpVisitor,
 
 
     conditions.IfStatementVisitor,
